chore: moved functions to make code more concise
feat: initialized components
