+++ middleend/src/ir/helper/types.rs
@@ -67,7 +67,6 @@ impl SlynxIR {
             Value::Instruction(instr) => self.get_type_of_instruction(instr.clone(), temp),
             Value::LabelArg(idx) => self.get_label(temp.current_label()).arguments()[*idx],
             Value::Slot(v) => self.get_slot_type(v.clone()),
-            Value::StructLiteral(ir, _) => *ir,
         }
     }
 
@@ -149,30 +148,6 @@ impl SlynxIR {
         let ty = self.types.get_function_type_mut(func_tyid);
         ty.insert_arg_types(&extended_args);
         ty.set_return_type(ret);
-        Ok(())
-    }
-    /// Inserts the contents of the provided `decl` into an IR struct type asserting it's a slynx component. This is made because component can be lowered to the equivalent of a struct with methods, thus 'classes'.
-    /// The thing is that this is made interanlly with the minimum of abstraction as possible, so it becomes a struct and the components as well as methods are inserted directly into the struct as fields
-    pub(crate) fn insert_component_fields_for(
-        &mut self,
-        decl: TypeId,
-        temp: &mut TempIRData,
-        tys: &TypesModule,
-    ) -> Result<(), IRError> {
-        let component_type = self.get_ir_type(&decl, temp, tys)?;
-        let IRType::Component(cid) = self.types.get_type(component_type) else {
-            unreachable!("Something errored that type of component simply isnt Component on ir");
-        };
-
-        let Some(HirType::Component { props: ty_props }) = tys.get_component(&decl) else {
-            unreachable!("{:?} should map to an Component, but it doesn't", decl);
-        };
-
-        for (_, _, prop) in ty_props {
-            let ty = self.get_ir_type(prop, temp, tys)?;
-            let comp_ty = self.types.get_component_type_mut(cid);
-            comp_ty.insert_field(ty);
-        }
         Ok(())
     }
 }
